Strategy Patterns Added
I've refactored a good chunk of the HandAnalysis class to instead
iterate through a Table held list of card rules.

This should allow for extensibility with custom rulesets defined by the
users in Poker and for the additional of new card games if so desired in
the future under a common architecture.

The Unit Tests covering HandAnalysis have been re-factored as needed in
response to these changes.